Crate icu_segmenter_data

Data for the icu_segmenter crate

This data was generated with CLDR version 48.2.1, ICU version release-78.1rc, and LSTM segmenter version v0.1.0.

Macros

impl_segmenter_break_grapheme_cluster_v1

Implement DataProvider<SegmenterBreakGraphemeClusterV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_grapheme_cluster_v2

Implement DataProvider<SegmenterBreakGraphemeClusterV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_line_override_v2

Implement DataProvider<SegmenterBreakLineOverrideV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_line_v1

Implement DataProvider<SegmenterBreakLineV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_line_v2

Implement DataProvider<SegmenterBreakLineV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_line_v3

Implement DataProvider<SegmenterBreakLineV3> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_sentence_override_v1

Implement DataProvider<SegmenterBreakSentenceOverrideV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_sentence_override_v2

Implement DataProvider<SegmenterBreakSentenceOverrideV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_sentence_v1

Implement DataProvider<SegmenterBreakSentenceV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_sentence_v2

Implement DataProvider<SegmenterBreakSentenceV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_word_override_v1

Implement DataProvider<SegmenterBreakWordOverrideV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_word_v1

Implement DataProvider<SegmenterBreakWordV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_break_word_v2

Implement DataProvider<SegmenterBreakWordV2> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_dictionary_auto_v1

Implement DataProvider<SegmenterDictionaryAutoV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_dictionary_extended_v1

Implement DataProvider<SegmenterDictionaryExtendedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_lstm_auto_v1

Implement DataProvider<SegmenterLstmAutoV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_unihan_radical_v1

Implement DataProvider<SegmenterUnihanRadicalV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.