Crate icu_segmenter_data

Data for the icu_segmenter crate

Macros

impl_segmenter_dictionary_w_auto_v1

Implement DataProvider<DictionaryForWordOnlyAutoV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_dictionary_wl_ext_v1

Implement DataProvider<DictionaryForWordLineExtendedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_grapheme_v1

Implement DataProvider<GraphemeClusterBreakDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_line_v1

Implement DataProvider<LineBreakDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_lstm_wl_auto_v1

Implement DataProvider<LstmForWordLineAutoV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_sentence_v1

Implement DataProvider<SentenceBreakDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_segmenter_word_v1

Implement DataProvider<WordBreakDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.